Delete old view files
