--- src/app/lang_old/new/page.tsx
@@ -1,11 +1,0 @@
-import { NewConlangForm } from "~/app/dashboard/_components/new-conlang-form";
-
-function NewConlangPage() {
-  return (
-    <div className="mt-8 flex flex-col items-center justify-center md:mt-24">
-      <NewConlangForm />
-    </div>
-  );
-}
-
-export default NewConlangPage;
