+++ src/hooks/data/useLexicalCategories.ts
@@ -2,7 +2,8 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import {
   createLexicalCategory,
   type CreateLexicalCategory,
-} from "~/app/lang_old/[id]/_actions/word";
+} from "~/server/actions/lexical-category";
+
 import { type LexicalCategory } from "~/types/word";
 
 function getLexicalCategories(conlangId: number) {
